requirements.txt
fastapi
uvicorn
celery
redis
sqlalchemy
psycopg2-binary
rdkit
Biopython
redis